requirements.txt
# Python 3.11+ compatible

# Core Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Deep Learning (CPU-only version for Ubuntu without GPU)
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# Computer Vision
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78  # For servers without display

# Data Processing
numpy==1.24.3
Pillow==10.1.0

# Security & Encryption
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Database
# SQLite is built into Python, no additional package needed

# Additional dependencies that might be needed on Ubuntu
# Uncomment if you encounter issues:
# gunicorn==21.2.0  # For production deployment
# python-magic==0.4.27  # For file type detection
